Moved DriveDirection and added speed argument
Moved the DriveDirection command from the autonomous package to the
drive package. Also made the command accept a speed.
