+++ src/org/usfirst/frc/team166/robot/commands/drive/DriveDirection.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team166.robot.commands.autonomous;
+package org.usfirst.frc.team166.robot.commands.drive;
 
 import edu.wpi.first.wpilibj.command.Command;
